* origin/master: (74 commits)
  Fix: removed VPN specific classes from container
  Fix: Subscription receipt input stream leak
  Replace subscription interval signal with GCD
  Refactored subscription update from server
  Fix: zombie threads for streaming subscription
  Rewrote subsription checking with ReactiveObjC
  Wrapped subscription and authorizations dictionary
  Added onFinished callback to RetryOperation
  Code refactoring
  Abstracted operation that retry with time interval
  Fix: showing homepage after downgrading connection
  Fix: VPN reassserting after restarting tunnel
  Added 'Authorizations' parameter to getPsiphonConfig
  Subscription - switch to using a signed authorization token - WIP
  Subscription - switch to using a signed authorization token - WIP
  Fix: use same logic for killing extension
  Fix: semantic error
  Updated PsiphonClientCommonLibrary and translations. Added Bengali.
  Subscription check in extension - WIP
  Newest PsiphonClientCommonLibrary with minor bug fixes
  ...

+++ PsiphonVPN/PacketTunnelUtils.m
@@ -45,7 +45,7 @@
         case NEProviderStopReasonConnectionFailed: return @"the connection failed";
     }
 
-    return [NSString stringWithFormat:@"unknown stop reason (%ld)", stopReason];
+    return [NSString stringWithFormat:@"unknown stop reason (%ld)", (long)stopReason];
 }
 
 @end
